requirements.txt
PyQt5==5.15.7
openpyxl==3.1.5
olefile==0.46
python-docx==0.8.11
Pillow==12.2.0
PyYAML==6.0.2
